The user side now has 95% of its controller handlers passing data through, including the graphs and tables. Again, these will need to be extracted from their current positions into either entirely new classes or functions since they are very very messy. No error handling has been performed so it will error out or not display warnings if bad info has been gathered. The mocksocket is not checking the token and is accepting all requests.

+++ src/Controllers/Backend/NetworkObjects/Trade.java
@@ -8,12 +8,12 @@ public class Trade implements Serializable, Comparable<Trade> {
     Integer TradeID;
     String AssetName;
     Integer AssetQuantity;
-    Float AssetPrice;
+    Double AssetPrice;
     String BuyerOrgName;
     String SellerOrgName;
     Date TradeDateMilSecs;
 
-    public Trade(Integer tradeID, String assetName, Integer assetQuantity, Float assetPrice, String buyerOrgName, String sellerOrgName, Date tradeDateMilSecs) {
+    public Trade(Integer tradeID, String assetName, Integer assetQuantity, Double assetPrice, String buyerOrgName, String sellerOrgName, Date tradeDateMilSecs) {
         TradeID = tradeID;
         AssetName = assetName;
         AssetQuantity = assetQuantity;
@@ -41,7 +41,7 @@ public class Trade implements Serializable, Comparable<Trade> {
         return AssetQuantity;
     }
 
-    public Float getAssetPrice() {
+    public Double getAssetPrice() {
         return AssetPrice;
     }
